Own file for SynchronisedChannelManager.
This splits up the gamechannel/channelmanager file into two, moving the
threading bits with SynchronisedChannelManager to a new syncmanager file.

The core channelmanager file now contains no longer any threading code
or other "complex" dependencies.
