+++ gamechannel/channelmanager_tests.cpp
@@ -17,8 +17,6 @@
 
 #include <glog/logging.h>
 
-#include <thread>
-
 using google::protobuf::TextFormat;
 using google::protobuf::util::MessageDifferencer;
 using testing::_;
@@ -27,6 +25,8 @@ using testing::Truly;
 
 namespace xaya
 {
+
+/* ************************************************************************** */
 
 proto::StateProof
 ValidProof (const std::string& state)
@@ -43,7 +43,8 @@ ValidProof (const std::string& state)
 ChannelManagerTestFixture::ChannelManagerTestFixture ()
   : cm(game.rules, game.channel,
        verifier, signer,
-       channelId, "player")
+       channelId, "player"),
+    onChain("game id", cm.GetChannelId (), "player", txSender, game.channel)
 {
   CHECK (TextFormat::ParseFromString (R"(
     participants:
@@ -63,6 +64,8 @@ ChannelManagerTestFixture::ChannelManagerTestFixture ()
 
   signer.SetAddress ("my addr");
   EXPECT_CALL (signer, SignMessage (_)).WillRepeatedly (Return ("sgn"));
+
+  cm.SetMoveSender (onChain);
 }
 
 void
@@ -85,39 +88,21 @@ ChannelManagerTestFixture::GetLatestState () const
   return UnverifiedProofEndState (cm.boardStates.GetStateProof ());
 }
 
+/* ************************************************************************** */
+
 namespace
 {
-
-class MockOffChainBroadcast : public OffChainBroadcast
-{
-
-public:
-
-  MockOffChainBroadcast (const uint256& i)
-    : OffChainBroadcast(i)
-  {
-    /* Expect no calls by default.  */
-    EXPECT_CALL (*this, SendMessage (_)).Times (0);
-  }
-
-  MOCK_METHOD1 (SendMessage, void (const std::string& msg));
-
-};
 
 class ChannelManagerTests : public ChannelManagerTestFixture
 {
 
 protected:
 
-  MockTransactionSender txSender;
-  MoveSender onChain;
   MockOffChainBroadcast offChain;
 
   ChannelManagerTests ()
-    : onChain("game id", channelId, "player", txSender, game.channel),
-      offChain(cm.GetChannelId ())
+    : offChain(cm.GetChannelId ())
   {
-    cm.SetMoveSender (onChain);
     cm.SetOffChainBroadcast (offChain);
   }
 
@@ -730,176 +715,6 @@ TEST_F (ChannelToJsonTests, PendingResolution)
   auto expected = Json::Value (Json::objectValue);
   expected["resolution"] = txid.ToHex ();
   EXPECT_EQ (cm.ToJson ()["pending"], expected);
-}
-
-/* ************************************************************************** */
-
-class WaitForChangeTests : public ChannelManagerTests
-{
-
-private:
-
-  /** The thread that is used to call WaitForChange.  */
-  std::unique_ptr<std::thread> waiter;
-
-  /** Set to true while the thread is actively waiting.  */
-  bool waiting;
-
-  /** Lock for waiting.  */
-  mutable std::mutex mut;
-
-  /** The JSON value returned from WaitForChange.  */
-  Json::Value returnedJson;
-
-protected:
-
-  /** Our synchronised manager for waiting.  */
-  SynchronisedChannelManager scm;
-
-  WaitForChangeTests ()
-    : scm(cm)
-  {}
-
-  /**
-   * Calls WaitForChange on a newly started thread.
-   */
-  void
-  CallWaitForChange (
-      int known = SynchronisedChannelManager::WAITFORCHANGE_ALWAYS_BLOCK)
-  {
-    CHECK (waiter == nullptr);
-    waiter = std::make_unique<std::thread> ([this, known] ()
-      {
-        LOG (INFO) << "Calling WaitForChange...";
-        {
-          std::lock_guard<std::mutex> lock(mut);
-          waiting = true;
-        }
-        returnedJson = scm.WaitForChange (known);
-        {
-          std::lock_guard<std::mutex> lock(mut);
-          waiting = false;
-        }
-        LOG (INFO) << "WaitForChange returned";
-      });
-
-    /* Make sure the thread had time to start and make the call.  */
-    SleepSome ();
-  }
-
-  /**
-   * Waits for the waiter thread to return and checks that the JSON value
-   * from it matches the then-correct ToJson output.  Also expects that the
-   * thread is finished "soon" (rather than timeout later).
-   */
-  void
-  JoinWaiter ()
-  {
-    CHECK (waiter != nullptr);
-
-    SleepSome ();
-    EXPECT_FALSE (IsWaiting ());
-
-    LOG (INFO) << "Joining the waiter thread...";
-    waiter->join ();
-    LOG (INFO) << "Waiter thread finished";
-    waiter.reset ();
-    ASSERT_EQ (returnedJson, cm.ToJson ());
-  }
-
-  /**
-   * Returns true if the thread is currently waiting.
-   */
-  bool
-  IsWaiting () const
-  {
-    CHECK (waiter != nullptr);
-
-    std::lock_guard<std::mutex> lock(mut);
-    return waiting;
-  }
-
-};
-
-TEST_F (WaitForChangeTests, OnChain)
-{
-  CallWaitForChange ();
-  ProcessOnChain ("0 0", ValidProof ("10 5"), 0);
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, OnChainNonExistant)
-{
-  CallWaitForChange ();
-  ProcessOnChainNonExistant ();
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, OffChain)
-{
-  ProcessOnChain ("0 0", ValidProof ("10 5"), 0);
-
-  CallWaitForChange ();
-  cm.ProcessOffChain ("", ValidProof ("12 6"));
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, OffChainNoChange)
-{
-  ProcessOnChain ("0 0", ValidProof ("10 5"), 0);
-
-  CallWaitForChange ();
-  cm.ProcessOffChain ("", ValidProof ("10 5"));
-
-  SleepSome ();
-  EXPECT_TRUE (IsWaiting ());
-
-  scm.StopUpdates ();
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, LocalMove)
-{
-  ExpectOneBroadcast ("11 6");
-  ProcessOnChain ("0 0", ValidProof ("10 5"), 0);
-
-  CallWaitForChange ();
-  cm.ProcessLocalMove ("1");
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, WhenStopped)
-{
-  scm.StopUpdates ();
-  CallWaitForChange ();
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, StopNotifies)
-{
-  CallWaitForChange ();
-  scm.StopUpdates ();
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, OutdatedKnownVersion)
-{
-  const int known = cm.ToJson ()["version"].asInt ();
-  ProcessOnChain ("0 0", ValidProof ("10 5"), 0);
-  CallWaitForChange (known);
-  JoinWaiter ();
-}
-
-TEST_F (WaitForChangeTests, UpToDateKnownVersion)
-{
-  const int known = cm.ToJson ()["version"].asInt ();
-  CallWaitForChange (known);
-
-  SleepSome ();
-  EXPECT_TRUE (IsWaiting ());
-
-  ProcessOnChain ("0 0", ValidProof ("10 5"), 0);
-  JoinWaiter ();
 }
 
 /* ************************************************************************** */
